backoffice - crud city in progress

+++ src/Controller/Back/CityController.php
@@ -5,6 +5,8 @@ namespace App\Controller\Back;
 use App\Entity\City;
 use App\Form\Back\CityType;
 use App\Repository\CityRepository;
+use DateTime;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -16,14 +18,20 @@ use Symfony\Component\Routing\Annotation\Route;
 class CityController extends AbstractController
 {
     /**
+     * Cities List
+     * 
      * @Route("/", name="index", methods={"GET"})
      * 
      * @ IsGranted("ROLE_ADMIN")
      */
-    public function index(CityRepository $cityRepository): Response
+    public function index(CityRepository $cityRepository, PaginatorInterface $paginatorInterface, Request $request): Response
     {
+        $cities = $cityRepository->findCitiesList();
+        $cities = $paginatorInterface->paginate($cities, 
+        $request->query->getInt('page', 1),6);
+
         return $this->render('back/city/index.html.twig', [
-            'cities' => $cityRepository->findAll(),
+            'cities' => $cities,
         ]);
     }
 
@@ -38,6 +46,11 @@ class CityController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $cityRepository->add($city, true);
+
+            $this->addFlash(
+                'success',
+                'La ville a bien été ajoutée'
+            );
 
             return $this->redirectToRoute('app_back_city_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -55,6 +68,7 @@ class CityController extends AbstractController
     {
         return $this->render('back/city/show.html.twig', [
             'city' => $city,
+            
         ]);
     }
 
@@ -67,7 +81,10 @@ class CityController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $city->setUpdatedAt(new DateTime("now"));
             $cityRepository->add($city, true);
+
+    
 
             return $this->redirectToRoute('app_back_city_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -83,6 +100,7 @@ class CityController extends AbstractController
      */
     public function delete(Request $request, City $city, CityRepository $cityRepository): Response
     {
+        if ($city === null){throw $this->createNotFoundException("Cette ville n'existe pas");}
         if ($this->isCsrfTokenValid('delete'.$city->getId(), $request->request->get('_token'))) {
             $cityRepository->remove($city, true);
         }
